pyproject.toml
[tool.poetry]
name = "clean-arch-python"
version = "0.1.0"
description = ""
authors = ["Cleyson Lima <cleysonph@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.2.2"
fastapi = "^0.79.1"
uvicorn = {extras = ["standard"], version = "^0.18.2"}
mysql-connector-python = "^8.0.30"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
